+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [unreleased]
+- No significant changes.
+
+## [0.4.0] - 2017-06-26
 - [breaking] JSON 출력 결과를 CamelCase 에서 underscore 변경
 - [breaking] Naver::Papago.romanization 리턴값의 key 변경. s_first_name -> first_name,
   a_items -> items
@@ -24,4 +27,5 @@ All notable changes to this project will be documented in this file.
 [0.1.0]: https://github.com/kimsuelim/naver-sdk-ruby/releases/tag/v0.1.0
 [0.2.0]: https://github.com/kimsuelim/naver-sdk-ruby/compare/v0.1.0...v0.2.0
 [0.3.0]: https://github.com/kimsuelim/naver-sdk-ruby/compare/v0.2.0...v0.3.0
-[unreleased]: https://github.com/kimsuelim/naver-sdk-ruby/compare/v0.3.0...HEAD
+[0.4.0]: https://github.com/kimsuelim/naver-sdk-ruby/compare/v0.3.0...v0.4.0
+[unreleased]: https://github.com/kimsuelim/naver-sdk-ruby/compare/v0.4.0...HEAD